Introduce ActionSelectionPolicy to refactor Q_learning
Implement e-greedy and softmax/boltzmann policies as
separate subclasses of ActionSelectionPolicy interface.
This approach will be used to refactor Q_learning and
potentially implement alternative policies without needing
to add if-statements in the agents' select_action methods

This implementation will be tested as a whole in agents' tests
without mocking the action selection policies.